Cleanup unused code and modernize
