+++ templates/index.html
@@ -131,7 +131,7 @@
       </div>
       <div class="modal-body">
         <div id="modal-effect-bg" class="particle-overlay"></div>
-        <img id="modal-img" src="" width="64" height="64" alt="">
+        <img id="modal-img" src="" width="64" height="64" alt="" loading="lazy">
         <div id="modal-details"></div>
       </div>
     </dialog>
@@ -141,11 +141,11 @@
       <p class="attribution">
         Pricing and particles provided by
         <a href="https://backpack.tf" target="_blank" rel="noopener" class="bptf-link">
-          Backpack.tf <img src="/static/images/logos/bptf_small.PNG" alt="Backpack.tf logo" class="footer-icon">
+          Backpack.tf <img src="/static/images/logos/bptf_small.PNG" alt="Backpack.tf logo" class="footer-icon" loading="lazy">
         </a>
         <br />
         Simplified Schemas provided by
-        <a href="https://autobot.tf/" target="_blank" rel="noopener" class="schema-link">Autobot.tf <img src="/static/images/logos/tf2autobot.png" alt="Autobot.tf logo" class="footer-icon"></a>
+        <a href="https://autobot.tf/" target="_blank" rel="noopener" class="schema-link">Autobot.tf <img src="/static/images/logos/tf2autobot.png" alt="Autobot.tf logo" class="footer-icon" loading="lazy"></a>
       </p>
 
       <p class="disclaimer">
