+++ utils/steam_api_client.py
@@ -3,7 +3,6 @@ from typing import Any, Dict, Iterator, List, Tuple
 
 import logging
 import httpx
-import requests
 import re
 from dotenv import load_dotenv
 
@@ -29,20 +28,33 @@ def _chunks(seq: List[str], size: int) -> Iterator[List[str]]:
         yield seq[i : i + size]
 
 
-def get_player_summaries(steamids: List[str]) -> List[Dict[str, Any]]:
-    """Return player summary data for all provided SteamIDs."""
-    results: List[Dict[str, Any]] = []
-    for chunk in _chunks(steamids, 100):
-        key = _require_key()
-        url = (
-            "https://api.steampowered.com/ISteamUser/GetPlayerSummaries/v2/"
-            f"?key={key}&steamids={','.join(chunk)}"
-        )
-        r = requests.get(url, timeout=10)
-        r.raise_for_status()
-        players = r.json().get("response", {}).get("players", [])
-        results.extend(players)
-    return results
+# ---------------------------------------------------------------------------
+# Steam ID parsing helpers
+
+STEAMID2_RE = re.compile(r"^STEAM_0:[01]:\d+$")
+STEAMID3_RE = re.compile(r"^\[U:1:\d+\]$")
+STEAMID64_RE = re.compile(r"^\d{17}$")
+
+
+def extract_steam_ids(raw_text: str) -> List[str]:
+    """Return unique Steam ID tokens found in ``raw_text``."""
+
+    tokens = re.split(r"\s+", raw_text.strip())
+    ids: List[str] = []
+    seen: set[str] = set()
+
+    for token in tokens:
+        if not token:
+            continue
+        if (
+            STEAMID2_RE.fullmatch(token)
+            or STEAMID3_RE.fullmatch(token)
+            or STEAMID64_RE.fullmatch(token)
+        ):
+            if token not in seen:
+                seen.add(token)
+                ids.append(token)
+    return ids
 
 
 async def get_player_summaries_async(steamids: List[str]) -> List[Dict[str, Any]]:
@@ -74,73 +86,6 @@ async def get_player_summaries_async(steamids: List[str]) -> List[Dict[str, Any]
                 players = []
             results.extend(players)
     return results
-
-
-def get_inventories(steamids: List[str]) -> Dict[str, Any]:
-    """Fetch TF2 inventories for each user via the Steam Web API."""
-    results: Dict[str, Any] = {}
-    headers = {"User-Agent": "Mozilla/5.0"}
-    for chunk in _chunks(steamids, 20):
-        for sid in chunk:
-            key = _require_key()
-            url = (
-                "https://api.steampowered.com/IEconItems_440/GetPlayerItems/v0001/"
-                f"?key={key}&steamid={sid}"
-            )
-            r = requests.get(url, headers=headers, timeout=10)
-            r.raise_for_status()
-            results[sid] = r.json().get("result", {})
-    return results
-
-
-def fetch_inventory(steamid: str) -> Tuple[str, Dict[str, Any]]:
-    """Fetch a user's inventory and classify visibility."""
-
-    headers = {"User-Agent": "Mozilla/5.0"}
-    key = _require_key()
-    url = (
-        "https://api.steampowered.com/IEconItems_440/GetPlayerItems/v0001/"
-        f"?key={key}&steamid={steamid}"
-    )
-
-    try:
-        resp = requests.get(url, headers=headers, timeout=20)
-    except requests.RequestException:
-        logger.info("Inventory %s: Fetch Failed", steamid)
-        return "failed", {}
-
-    if resp.status_code in (400, 403):
-        logger.info("Inventory %s: Private", steamid)
-        return "private", {}
-
-    if resp.status_code != 200:
-        logger.info("Inventory %s: HTTP %s", steamid, resp.status_code)
-        return "failed", {}
-
-    try:
-        data = resp.json()
-    except ValueError:
-        logger.info("Inventory %s: invalid JSON", steamid)
-        return "failed", {}
-
-    result = data.get("result")
-    if not isinstance(result, dict):
-        logger.info("Inventory %s: Private", steamid)
-        return "private", {}
-    status_code = result.get("status")
-    items = result.get("items") or []
-
-    if status_code == 1:
-        if items:
-            logger.info(
-                "Inventory %s: Public and Parsed (%s items)", steamid, len(items)
-            )
-            return "parsed", result
-        logger.info("Inventory %s: Public but Empty", steamid)
-        return "incomplete", result
-
-    logger.info("Inventory %s: Private", steamid)
-    return "private", result
 
 
 async def fetch_inventory_async(steamid: str) -> Tuple[str, Dict[str, Any]]:
@@ -214,25 +159,6 @@ def convert_to_steam64(id_str: str) -> str:
         return str(z + 76561197960265728)
 
     raise ValueError(f"Invalid Steam ID format: {id_str}")
-
-
-def get_tf2_playtime_hours(steamid: str) -> float:
-    """Return TF2 playtime in hours for a Steam user."""
-    url = "https://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/"
-    key = _require_key()
-    params = {
-        "key": key,
-        "steamid": steamid,
-        "include_played_free_games": 1,
-        "format": "json",
-    }
-    r = requests.get(url, params=params, timeout=10)
-    r.raise_for_status()
-    data = r.json().get("response", {})
-    for game in data.get("games", []):
-        if game.get("appid") == 440:
-            return game.get("playtime_forever", 0) / 60.0
-    return 0.0
 
 
 async def get_tf2_playtime_hours_async(steamid: str) -> float:
